// src/alerts/types/Alert.ts
export type AlertLevelParam = string | string[];

// Alert status search parameter types.
export type AlertStatusParam = string | string[];

// Alert list search parameters.
export interface AlertSearchParams extends AlertTimeSearchParams {
  // Category to filter by.
  categories?: number;
  // Alert level to search for.
  level?: AlertLevelParam;
  // Alert status to search for.
  status?: AlertStatusParam;
  // Assigned user to search for.
  assigned_user?: number;
  // Distillery the alert came from.
  collection?: number;
  // String content the alert data contains.
  content?: string;
  // Limit of alerts to return.
  limit?: number;
  // How many alerts to offset.
  offset?: number;
}

// Alert time search parameters.
export interface AlertTimeSearchParams {
  // Time after to search for alerts.
  after?: string;
  // Time before to search for alerts.
  before?: string;
}
// ...
export interface Category {
  // ID of the category.
  id: number;
  // Name of the category.
  name: string;
}

// Normalized category entities.
export interface CategoryEntities {
  // Normalized category objects.
  categories: NormalizedEntity<Category>;
}

// List of normalized category objects.
export type NormalizedCategoryList = NormalizedList<number, CategoryEntities>;